requirements.txt
flask>=2.3.0
flask-cors>=4.0.0
kafka-python>=2.0.2
redis>=4.5.0
spacy>=3.7.0
transformers>=4.30.0
feedparser>=6.0.0
requests>=2.31.0
beautifulsoup4>=4.12.0
schedule>=1.2.0
asyncio>=3.4.3
aiohttp>=3.8.0
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
sentence-transformers>=2.2.0
python-dotenv>=1.0.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
pymilvus>=2.3.0
elasticsearch>=8.0.0

